requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
supabase>=2.3.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyyaml>=6.0.0
loguru>=0.7.0
tqdm>=4.65.0
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
faiss-cpu>=1.7.0
